Move cache package to data package
