+++ kt/plugins/godot-intellij-plugin/src/main/kotlin/godot/intellij/plugin/data/cache/RegisteredClassNameCache.kt
@@ -1,4 +1,4 @@
-package godot.intellij.plugin.cache
+package godot.intellij.plugin.data.cache
 
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
